fix(react): clear remaining react-doctor warnings in app/website/mobile
- app: split CloudVideoInputNode into VideoIcon + UrlInputSection siblings
  (no-multi-comp); main file is now the sole component
- website: inline heroProof/heroVisual JSX at PageLayout prop sites in
  product-page + use-cases-content (jsx-no-jsx-as-prop)
- mobile/analytics: move early returns above refreshControl, inline
  RefreshControl at prop site (rerender-memo-before-early-return)
- mobile/approvals: drop refreshControl useMemo, inline RefreshControl at
  FlashList prop site (rerender-memo-before-early-return)

website/desktop/mobile/extension now report 0 react-doctor issues.
Remaining app warnings are classified false-positives + maintainability
risk-skips (React Query status bridges, RSC jsx-as-prop, giant-component).
0 errors; CI react-doctor gate stays green.
