+++ apps/app/src/features/workflows/nodes/input/VideoIcon.tsx
@@ -1,0 +1,20 @@
+'use client';
+
+export function VideoIcon({
+  className = 'h-4 w-4',
+}: {
+  className?: string;
+}): React.JSX.Element {
+  return (
+    <svg
+      className={className}
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth="2"
+    >
+      <polygon points="23 7 16 12 23 17 23 7" />
+      <rect x="1" y="5" width="15" height="14" rx="2" ry="2" />
+    </svg>
+  );
+}
