+++ apps/mobile/app/app/(protected)/analytics.tsx
@@ -1,4 +1,3 @@
-import { useMemo } from 'react';
 import {
   RefreshControl,
   ScrollView,
@@ -191,17 +190,6 @@ function EngagementBreakdown({
 export default function AnalyticsScreen() {
   const { data, isLoading, error, refetch } = useAnalytics();
 
-  const refreshControl = useMemo(
-    () => (
-      <RefreshControl
-        refreshing={isLoading}
-        onRefresh={refetch}
-        tintColor={colors.indigo}
-      />
-    ),
-    [isLoading, refetch],
-  );
-
   if (isLoading && !data.overview) {
     return (
       <LoadingScreen message="Loading analytics..." color={colors.indigo} />
@@ -224,7 +212,13 @@ export default function AnalyticsScreen() {
     <ScrollView
       style={styles.container}
       contentContainerStyle={styles.contentContainer}
-      refreshControl={refreshControl}
+      refreshControl={
+        <RefreshControl
+          refreshing={isLoading}
+          onRefresh={refetch}
+          tintColor={colors.indigo}
+        />
+      }
     >
       <View style={styles.header}>
         <Text style={styles.headerTitle}>Analytics</Text>
